requirements.txt
Django>=4.2.8
django-environ
mysqlclient
mysql-connector-python
django-decorator-include
setuptools
